fix(rogue): add missing autoFocus and rename utilities
